Added create subject to remote functions and fix transition state from domain to subject
