+++ src/routes/graph-editor/graphs/[graphid=int]/subjects/subjects.remote.ts
@@ -3,7 +3,7 @@ import { getUser } from '$lib/server/actions/Users';
 import prisma from '$lib/server/db/prisma';
 import { whereHasGraphCoursePermission } from '$lib/server/permissions';
 import { svelteError } from '$lib/utils/setError';
-import { createSubjectSchema } from '$lib/valibot/subjectSchema';
+import { createSubjectSchema, subjectSchema } from '$lib/valibot/subjectSchema';
 
 export const createSubject = form(createSubjectSchema, async ({ graphId, name, domainId }) => {
 	const user = await getUser(getRequestEvent());
@@ -31,6 +31,27 @@ export const createSubject = form(createSubjectSchema, async ({ graphId, name, d
 						domainId: domainId > 0 ? domainId : null
 					}
 				}
+			}
+		});
+		return { success: true };
+	} catch (e) {
+		svelteError(e);
+	}
+});
+
+export const changeSubject = form(subjectSchema, async ({ graphId, subjectId, name, domainId }) => {
+	const user = await getUser(getRequestEvent());
+
+	try {
+		await prisma.subject.update({
+			where: {
+				id: subjectId,
+				graphId: graphId,
+				...whereHasGraphCoursePermission(user, 'CourseAdminEditorORProgramAdminEditor')
+			},
+			data: {
+				name: name,
+				domainId: domainId > 0 ? domainId : null
 			}
 		});
 		return { success: true };
